A hardware manager call for a full sync before shutdown
This is largely required for the future lockdown command but can also be
used before the normal shutdown, especially in the sync command which is
currently used before an out-of-band shutdown command is issued.

In addition to a plain sync, the new command also tells the kernel to
drop its cached and issues a low-level sync command to each block
device.

Partial-Bug: #2077432
Change-Id: I3fc87b20bc5387a466b24ebc19b9982e4e368d20