Vendor metrics library from Ironic-Lib & deprecate
We are phasing out use of ironic-lib, and as such are removing the
metrics module from it. However, due to it's requirement of having
a statsd instance on the same subnet as the agent and there being no
support for prometheus exporting of metrics from IPA, these metrics are
no longer valuable (in the agent).

We are vendoring the module for the deprecation in order to facilitate
its removal from ironic-lib.

Change-Id: Ie50e078bc3f78d65cfa53680dc4116d1119ce155